WIP - enable toggle b/w feature VCs; add parent class for features to house shared logic (ex: create and process orderID)
